fix types in the assertion

+++ span_test.go
@@ -33,8 +33,8 @@ func TestBasicSpan(t *testing.T) {
 	assert.NotEmpty(t, span.SpanID)
 	assert.NotEmpty(t, span.TraceID)
 	assert.NotEmpty(t, span.Timestamp)
-	assert.LessOrEqual(t, 10*time.Millisecond, span.Duration)
-	assert.LessOrEqual(t, span.Duration, elapsed)
+	assert.LessOrEqual(t, uint64(10*time.Millisecond), span.Duration)
+	assert.LessOrEqual(t, span.Duration, uint64(elapsed))
 	assert.Equal(t, "sdk", span.Name)
 
 	require.IsType(t, instana.SDKSpanData{}, span.Data)
